move device detection file under utils
