+++ unitylibs/utils/device-detection.js
